requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
python-dotenv>=1.0.0
Werkzeug>=3.0.1
selenium>=4.0.0
webdriver-manager>=4.0.0
pandas>=2.0.0
openpyxl>=3.0.0
xlrd>=2.0.0
pytz>=2023.3
psutil>=5.9.0
APScheduler>=3.10.4
requests>=2.31.0